[stevo]: combine generates, make it a trait
